requirements.txt
openai~=1.3.3
python-dotenv~=0.19.2
discord~=2.3.2